+++ agents-flex-core/src/main/java/com/agentsflex/core/chain/DataType.java
@@ -16,6 +16,7 @@
 package com.agentsflex.core.chain;
 
 public enum DataType {
+    Array("Array"),
     Object("Object"),
     String("String"),
     Number("Number"),