+++ agents-flex-llm/agents-flex-llm-openai/src/test/java/com/agentsflex/llm/openai/OpenAILlmTest.java
@@ -10,6 +10,8 @@ import com.agentsflex.core.prompt.ImagePrompt;
 import com.agentsflex.core.prompt.ToolPrompt;
 import org.junit.Test;
 
+import java.util.concurrent.TimeUnit;
+
 public class OpenAILlmTest {
 
     @Test(expected = LlmException.class)
@@ -116,5 +118,183 @@ public class OpenAILlmTest {
         } else {
             System.out.println(response);
         }
+    }
+
+    @Test()
+    public void testFunctionCalling3() throws InterruptedException {
+        OpenAILlmConfig config = new OpenAILlmConfig();
+        config.setDebug(true);
+        config.setEndpoint("https://ark.cn-beijing.volces.com");
+        config.setChatPath("/api/v3/chat/completions");
+        config.setModel("deepseek-v3-250324");
+        config.setApiKey("2d57a");
+
+        OpenAILlm llm = new OpenAILlm(config);
+
+        FunctionPrompt prompt = new FunctionPrompt("今天北京的天气怎么样", WeatherFunctions.class);
+        AiMessageResponse response = llm.chat(prompt);
+
+        if (response.isFunctionCall()) {
+            AiMessageResponse response1 = llm.chat(ToolPrompt.of(response));
+            System.out.println(response1.getMessage().getContent());
+        } else {
+            System.out.println(response);
+        }
+    }
+
+    @Test()
+    public void testFunctionCalling4() throws InterruptedException {
+        OpenAILlmConfig config = new OpenAILlmConfig();
+        config.setDebug(true);
+        config.setEndpoint("https://ark.cn-beijing.volces.com");
+        config.setChatPath("/api/v3/chat/completions");
+        config.setModel("deepseek-v3-250324");
+        config.setApiKey("2d57aa75");
+
+        OpenAILlm llm = new OpenAILlm(config);
+
+        FunctionPrompt prompt = new FunctionPrompt("今天北京的天气怎么样", WeatherFunctions.class);
+        llm.chatStream(prompt, new StreamResponseListener() {
+            @Override
+            public void onMessage(ChatContext context, AiMessageResponse response) {
+                System.out.println(" onMessage >>>>>" + response.isFunctionCall());
+            }
+        });
+
+        TimeUnit.SECONDS.sleep(5);
+    }
+
+    @Test()
+    public void testFunctionCalling44() throws InterruptedException {
+        OpenAILlmConfig config = new OpenAILlmConfig();
+        config.setDebug(true);
+        config.setEndpoint("https://ark.cn-beijing.volces.com");
+        config.setChatPath("/api/v3/chat/completions");
+        config.setModel("deepseek-v3-250324");
+        config.setApiKey("2d5");
+
+        OpenAILlm llm = new OpenAILlm(config);
+
+        FunctionPrompt prompt = new FunctionPrompt("北京和上海的天气怎么样", WeatherFunctions.class);
+        llm.chatStream(prompt, new StreamResponseListener() {
+            @Override
+            public void onMessage(ChatContext context, AiMessageResponse response) {
+                System.out.println(" onMessage >>>>>" + response.isFunctionCall());
+            }
+        });
+
+        TimeUnit.SECONDS.sleep(5);
+    }
+
+    @Test()
+    public void testFunctionCalling444() throws InterruptedException {
+        OpenAILlmConfig config = new OpenAILlmConfig();
+        config.setDebug(true);
+        config.setEndpoint("https://ark.cn-beijing.volces.com");
+        config.setChatPath("/api/v3/chat/completions");
+        config.setModel("deepseek-v3-250324");
+        config.setApiKey("2d57a");
+
+        OpenAILlm llm = new OpenAILlm(config);
+
+
+        FunctionPrompt prompt = new FunctionPrompt("北京和上海的天气怎么样", WeatherFunctions.class);
+        llm.chatStream(prompt, new StreamResponseListener() {
+            @Override
+            public void onMessage(ChatContext context, AiMessageResponse response) {
+//                System.out.println("onMessage >>>>>" + response);
+                if (response.isFunctionCall()) {
+                    System.out.println(":::::::: start....");
+                    llm.chatStream(ToolPrompt.of(response), new StreamResponseListener() {
+                        @Override
+                        public void onMessage(ChatContext context, AiMessageResponse response) {
+                            String msg = response.getMessage().getContent() != null ? response.getMessage().getContent() : response.getMessage().getReasoningContent();
+                            System.out.println(":::"+msg);
+                        }
+                    });
+                } else {
+                    String msg = response.getMessage().getContent() != null ? response.getMessage().getContent() : response.getMessage().getReasoningContent();
+                    System.out.println(">>>"+msg);
+                }
+            }
+        });
+
+        TimeUnit.SECONDS.sleep(25);
+
+    }
+
+
+    @Test()
+    public void testFunctionCalling5() throws InterruptedException {
+        OpenAILlmConfig config = new OpenAILlmConfig();
+        config.setDebug(true);
+        config.setEndpoint("https://ai.gitee.com");
+        config.setModel("Qwen3-32B");
+        config.setApiKey("PXW1");
+
+        OpenAILlm llm = new OpenAILlm(config);
+
+        FunctionPrompt prompt = new FunctionPrompt("北京和上海的天气怎么样", WeatherFunctions.class);
+        llm.chatStream(prompt, new StreamResponseListener() {
+            @Override
+            public void onMessage(ChatContext context, AiMessageResponse response) {
+                System.out.println("onMessage >>>>>" + response);
+            }
+        });
+
+        TimeUnit.SECONDS.sleep(5);
+    }
+
+
+    @Test()
+    public void testFunctionCalling55() throws InterruptedException {
+        OpenAILlmConfig config = new OpenAILlmConfig();
+        config.setDebug(true);
+        config.setEndpoint("https://ai.gitee.com");
+        config.setModel("Qwen3-32B");
+        config.setApiKey("PXW1");
+
+        OpenAILlm llm = new OpenAILlm(config);
+
+        FunctionPrompt prompt = new FunctionPrompt("/no_think 北京和上海的天气怎么样", WeatherFunctions.class);
+//        FunctionPrompt prompt = new FunctionPrompt("上海的天气怎么样", WeatherFunctions.class);
+        llm.chatStream(prompt, new StreamResponseListener() {
+            @Override
+            public void onMessage(ChatContext context, AiMessageResponse response) {
+//                System.out.println("onMessage >>>>>" + response);
+                if (response.isFunctionCall()) {
+                    System.out.println(":::::::: start....");
+                    llm.chatStream(ToolPrompt.of(response), new StreamResponseListener() {
+                        @Override
+                        public void onMessage(ChatContext context, AiMessageResponse response) {
+                            String msg = response.getMessage().getContent() != null ? response.getMessage().getContent() : response.getMessage().getReasoningContent();
+                            System.out.println(":::"+msg);
+                        }
+                    });
+                } else {
+                    String msg = response.getMessage().getContent() != null ? response.getMessage().getContent() : response.getMessage().getReasoningContent();
+                    System.out.println(">>>"+msg);
+                }
+            }
+        });
+
+        TimeUnit.SECONDS.sleep(25);
+    }
+
+    @Test()
+    public void testFunctionCalling6() throws InterruptedException {
+        OpenAILlmConfig config = new OpenAILlmConfig();
+        config.setDebug(true);
+        config.setEndpoint("https://ai.gitee.com");
+        config.setModel("Qwen3-32B");
+        config.setApiKey("PXW1");
+
+        OpenAILlm llm = new OpenAILlm(config);
+
+        FunctionPrompt prompt = new FunctionPrompt("/nothink 北京和上海的天气怎么样", WeatherFunctions.class);
+        AiMessageResponse response = llm.chat(prompt);
+
+        System.out.println(llm.chat(ToolPrompt.of(response)));
+
     }
 }